+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Improvements:
 * Core: upgraded Spoon
 * Core: new CKFinder license, see: http://www.fork-cms.com/blog/detail/new-ck-finder-license
 * Core: Include a non-official patch for CKeditor to fix an issue with the stylesheetparser on FF/Safari on Macs.
+* Core: utils.string.ucfirst added, it was missing in frontend. Tx to Jeroen Desloovere
 
 Bugfixes: